BE/ai-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
oracledb>=2.0.0
python-dotenv==1.0.0
pydantic>=2.7.4

langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-core>=0.3.0
langchain-community>=0.3.0
rank-bm25==0.2.2

chromadb==1.0.21
sentence-transformers>=2.3.0

openai>=1.40.0,<2.0.0

numpy>=1.26.0
requests==2.31.0
apscheduler==3.10.4
httpx>=0.27.0

# 허깅페이스에서 models--cross-encoder--ms-marco-MiniLM-L-6-v2 이걸 로컬에 다운받음. (reranker모델)